pyproject.toml
[project]
name = "langchain-yt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "google-genai>=1.68.0",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.13",
    "langchain-classic>=1.0.3",
    "langchain-core>=1.2.21",
    "langchain-google-genai>=4.2.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
